move source match termination to augen_agentchat
