--- python/packages/autogen-ext/src/autogen_ext/task/__init__.py
@@ -1,3 +1,0 @@
-from ._source_match_termination import SourceMatchTermination
-
-__all__ = ["SourceMatchTermination"]
